Make Card component reusable
