+++ src/components/docs/FAQ.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { css, withStyles } from 'withStyles';
+import { compose, withState, withHandlers } from 'recompose';
+import { Row, Col } from 'antd';
